HMCR-670

+++ api/Hmcr.Model/Dtos/ActivityCode/ActivityCodeCreateDto.cs
@@ -20,6 +20,6 @@ namespace Hmcr.Model.Dtos.ActivityCode
         public virtual IList<decimal> ServiceAreaNumbers { get; set; }
         public decimal? MinValue { get; set; }
         public decimal? MaxValue { get; set; }
-        public decimal? ReportingFrequency { get; set; }
+        public int? ReportingFrequency { get; set; }
     }
 }
